pyproject.toml
[tool.poetry]
name = "edonet"
version = "0.1.0"
description = "A minimal neural network using cupy."
authors = ["Edo van Veen <edo@vanveen.io>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
cupy-cuda12x = "^13.1.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"
pytest = "^8.1.1"
mypy = "^1.10.0"
matplotlib = "^3.8.4"
scikit-learn = "^1.4.2"
numpy = "^1.26.4"

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
